added rational number approximation to find FRAC/MOD for fractional N mode added greatest common divisor calculation to fine best combiation of N and R for integer N mode added printout of named register fields to with binary and hex printouts of whole registers

+++ Software/include/rational_approximation.h
@@ -3,11 +3,11 @@
 #define RATIONAL_APPROXIMATION_H
 
 typedef struct {
-    uint32_t numerator;
-    uint32_t denominator;
-    uint32_t iterations;   // Just for debugging
+    uint16_t numerator;
+    uint16_t denominator;
+    uint16_t iterations;   // Just for debugging
   } rational_t;
 
-rational_t rational_approximation(double target, uint32_t maxdenom);
+rational_t rational_approximation(float target, uint16_t maxdenom);
 void test_rational_approx();
 #endif
